ref(remote): commands declare a state instead of toggling
The act of toggling could cause a race condition where
someone intends to unmute, and then while that request is
in flight someone else already tried to unmute, causing
the unmute to actually become a mute. So instead be
explicit about what new state to set and no-op if already
in that state.