fix(meeting): workaround for flashing background
On my low end, beater laptop (happens to run
windows 10), when an active speaker change happens
the Spot application background may flash on screen
for less than a second. The issue seems to happen
within the iframe itself, outside of the scope of
Spot. As a workaround, give the wrapper around the
iframe a background color to at least mask whatever
background image or color Spot uses.