feat(join-code): custom component for code entry
The current version uses multiple input elements
and switches focus between them. In webviews the
blurring of an input causes the keyboard to be
dismissed and the next input not to be blurred.
The new version uses one input and captures
keyboard events to update the UI to avoid
switching focus on inputs.