fix(ultrasound): move back to playing from the web (#79)
This consolidates logic back into the web so the
ipad app can remain as logic free as possible.
This is possible because the iPad webview was
changed to support playing sound (ultrasound)
without user input. This also prevents the
passing of sensitive information through
the webview to the native code through post
message.

However, in moving ultrasound to web, a fork of
quietjs is being used which adds support for npm
and reduces script tags loads. The fork has been
in flight for almost 3 weeks with no end known
The rewrite of it to support npm and reduce script
tag loads is still in flight after almost 3 weeks
so this fork is being used to unblock testing.