fix(calendar): notify remotes of events after calendar change
The edge case is being connected to a calendar with events,
going into setup, changing calendars to one without any
events. No notification of the new events would be sent
out to remotes, leaving them to show events from the
previous calendar. This is because the known events
for the new calendar had not changed (going from empty
to empty) and a calendar fetch had been completed
for the old calendar. The fix is clearing the fetch
flag with the calendar change.