fix(logging): prevent double json stringification (#136)
Objects in logs were being logged as strings,
instead of key-accessible objects. The behavior
was intentional because of how jitsi-meet-logger
concatenates logs as one string, and converting
back to JSON could cause errors. However, this
produces the undesireable log format, so introduce
a context argument object for logging additional
details.