fix(reconnect): ensure calendar events are emitted
When reconnecting, Spot-TV creates a new XMPP
connection and sets a new join code. The setting
of a join code triggers a redux update, which
updates presence through a redux subscriber.
This ensure the new connection has the updated
presence. Use this mechanism to enusre the
new connection has updated events as well.