fix(screenshare): stop relying on screensharing boolean
There was a transitionary time when jitsi-meet only said
if screensharing was on or off and not the type of screensharing
that is currently enabled. During this time both the boolean
and screenshare type were used to determine the current spot
screenshare status. With the jitsi-meet changes deployed,
just rely on the screenshare type to know screenshare is on.