ref(css): use sass partials to avoid duplicate imports
sass will import a file for each @import declaration,
and multiple @import declarations for the same file
results in multiple copies of the file being bundled.
The workaround is make everything a partial that gets
blasted into a single file, in this case a new
index.scss file.

This change currently drops about .02mb (20kb) from
the bundle.