fix(reconnect): RCS should clean up own connection on disconnect
If a disconnect occurs, or a failed connection occurs,
it was possible to attempt the before the disconnect
logic finished. This could lead to a state where the
remote is connected to the service but the reference
to the xmpp connection was made null. To fix this,
ensure the connection is cleaned up and events of
a disconnect happen after cleanup.