fix(calendar): allow searching for room in google calendar
Google and Outlook truncate room list results to 100. If
a company has more than 100 rooms then not all rooms will
be returned. The Google API allows for querying for rooms,
which provides some search functionality, so that has been
introduced to get around the 100 room limit. Outlook does
not seem to provide such so it's going to have to survive
for a bit in its current state.

Note that this is a partial solution until the real
backend service is complete. Google's API querying matches
to the start of field values only, not in the middle,
and is case sensitive.