ref(RCS): move exchangeCode call to be part of RCS connect
Make exchangeCode call to be part of the connect, so that it's part of
the reconnect process.

This also fixes an issue with RCS running with no backend where Spot
Remote will fail to rejoin after the join code has been refresh. That's
because it was using roomInfo stored in this._options instead of
the join code.