ref: introduce remoteJoinCode
With the backend Spot Remote will use different join code that Spot TV.
After this commit a `joinCode` is the one used by the local instance to
connect to the service and establish pairing between Spot TV and Spot
Remote.

A remote pairing code is produced only by a Spot TV and displayed on
the screen, so that a user can use it to pair a Spot Remote.