feat(Spot Remote): store a permanent pairing code
If the backend grants a refresh token to the client it means that
the connection can be refresh indefinitely. In such case the join code
will be persisted and re-used next time the app is started.

Adds top level retry logic for permanently paired remotes to the connect
action (similar to Spot TV).