fix(rcs): pass onbeforeunload event to xmpp
JitsiConnection#disconnect takes in a browser
event object, which gets passed to its xmpp
connection. If that event is for a beforeunload
or unload event, the the disconnect is fired
synchronously to help ensure it completes.

This should help prevent jid conflicts
while developing and potentially with
upcoming work for selenium tests.