ref: do not clear token on NO_JWT and add REGISTRATION_LOST event
NO_JWT error is just a check on whether or not SpotBackendService has
a valid JWT, but this could be due to a bug in the code or response
structure change on the backend. This should not remove the refresh token
from the storage and the service should keep trying as update could fix
the issue.

Adds the REGISTRATION_LOST event to the backend service which in backend
service terms means disconnected.