feat(RCS): specify MUC JID resource name for Spot Remotes
Will use 'remote-perm' prefix for a permanent Spot Remote and
'remote-temp' for a temporary one. This will allow the backend to figure
out what's being connected to the MUC.

Example value for a permanent Spot Remote:

'remote-perm-12345678'

Example for a temporary Spot Remote:

'remote-temp-123455678'

The numbers at the end will be 8 random hex characters.