feat(SpotBackendService.test): add cases which cover JWT not updated
In case where the RemoteControlService is in the reconnect loop it does
not do anything with the results of 'register' method which may contain
the new JWT(if reconnecting after JWT expired). In that case the backend
is not emitting REGISTRATION_UPDATED events and JWT update is not
propagated to the Redux.