fix(RCS client): clear Spot TV state on disconnect
Fixes a case where Spot Remote will end up on not working remote control
screen if Spot TV leaves while the Remote is disconnected/reconnecting.
It will display "waiting for Spot TV to connect..." message instead.

A side effect of this change is that whenever RCS starts reconnecting
the client will no longer show the calendar page(or whatever's been left
opened) behind the reconnect overlay. In order to avoid that we'd have
to clear the TV state just before connect or use a hack, but that does
not sound correct.