ref(ultrasound): remove workaround of playing on interaction
Before it was known a mobile application was required,
ultrasound was built to be played from mobile safari,
which requires a user interaction before playing sound.
The workaround was to start playing sound on join code
submit, but that is not needed anymore as the react
native webview can allow sound to play without
interaction.