fix(RCS): emit JWT update after reconnect is completed
While the reconnect is in progress there are no listeners bound to
the backend service, but are re-added after the reconnect is completed.
It happens that the JWT can be refreshed during the reconnect and this
update will be missed.

This commit emits extra update event when the reconnect is done.
The reason for that is to keep the logic that removes the listeners
on disconnect. Also when connect/reconnect happens
the failure is handled on the promise chain instead of listeners.
Keeping the listeners could trigger double disconnect handing or other
weird things.

Maybe it's better to bind the listeners before connect and handle
updates/disconnect from there instead of promise chains?