ref: remove the reconnect logic from the remote control service
Remote control service will no longer deal with reconnects. From now on
the only reconnect logic will be in the actions which start
the connections.

This change also makes temporary Spot Remote drop the connection and go
back to the join code entry page if RCS disconnects.

Also there will be no reconnect overlay as it was displayed when RCS was
trying to reconnect. Now both TV and Remote will display the loading
screen while reconnecting.