feat(dial pad): pretty format and validate phone numbers (#560)
* feat(dial pad): pretty format and validate phone numbers

Use libphonenumber-js to format the displayed phone number as it's being
typed. Also validates if the phone number is valid, before calling
the onSubmit callback.

Uses 'US' as a default country code for formatting and validation if no
country code is entered manually.

Modifies StatelessDialPad to not pass the value and only call onSubmit
callback when submit action is executed. It is not supposed to track
the state anyway and it is confusing whether to use the value given on
submit or the value that's been submitted last in the on change
callback.