feat(dial pad): detect the country as phone number is being typed
Tracks the 'country' value exposed by the "as you type" object and
update the selection when another country is detected. Also adds
the + prefix for the country codes.

Note that it doesn't work reliably for countries where the prefix is
ambiguous. In such cases country is detected only once the number has
been fully typed.