feat(dial): country picker (#566)
* feat(dial): country picker

* ref(countriesInfo): add missing countries

Adds countries supported by the libphonenumber-js according to
the metadata source at:

https://raw.githubusercontent.com/googlei18n/libphonenumber/master/resources/PhoneNumberMetadata.xml

* feat(dial pad): detect the country as phone number is being typed

Tracks the 'country' value exposed by the "as you type" object and
update the selection when another country is detected. Also adds
the + prefix for the country codes.

Note that it doesn't work reliably for countries where the prefix is
ambiguous. In such cases country is detected only once the number has
been fully typed.

* ref(dial-pad): do not insert country code prefix into the phone number

There's no need to insert country code prefix if the country is set on
the "as you type" object. It will add it automatically to the E.164
phone number.

* fix dial pad tests

* fix log context

* fix eslint