fix(RCS): connection.passwordRequires is recoverable with backend
In the backend flow 'connection.passwordRequired' is recoverable error.
It can happen that device goes to sleep after verifying the token and
fetching room info then it wakes up and tries to use outdated token to
connect to xmpp. In this case it would fail with 'passwordRequired' and
what needs to happen is that the connection flow needs to start over and
refresh the token.

Another case would be when the XMPP part is misconfigured and not
accepting tokens provided by the backend, but this should not result
in lost pairing, but the client should keep on trying to connect in hope
the issue is resolved on the infrastructure side.

(cherry picked from commit e426d301af7197387466b35f1167257398ae4e4e)