feat(electron): add the machinery for reliable electron messaging
Adds functionality for sending message from spot-electron to
the spot-client counterpart. It allows for reliably sending messages
with making sure that nothing gets lost in case the spot-client has
crashed or got destroyed. In order to achieve that the spot-client must
ping spot-electron once it's bootstrap has completed. Then spot-electron
must watch for 'crashed' and 'destroyed' events on the web contents
object passed in as the sender of the 'spot-client/ready' ping message.