feat(electron): integrate 'electron-updater'
Refactors spot-client to provide an OK to update signal to the native
electron counterpart. It is okay to update if the current time falls
withing the preconfigure time update range and if the home screen is
displayed by the Spot TV. The spot-electron part will the check for
update, download and install by restarting the app.