feat(electron): update when the app starts
In order to prevent client being locked on a specific spot version,
check for the update and download on startup. This will update
the client on next restart or as soon as the spot-client gives a green
light using the OK to update message.

Also adjusts the Ok to update logic to allow an update if the client
does not send any message. It is expected from spot-client to send a not
OK to update message if it's in a meeting or other situation where it
would disrupt the current user interaction.