fix(RCS): leaked xmpp connection
The issue happens if the XMPP connection dies first and then the P2P
connection dies just after. In this case the silent reconnect will start
connection process, but as soon P2P dies it will drop the RCS connection
with 'no-active-signaling', but the silent reconnect process is still
active and will leak the connection if succeeds.

This commit adds extra 'reconnecting' getter and will let the reconnect
either fail or succeed if it's already in progress.