ref(RCS+P2P): initialize P2P instance early + do not kill when TV leaves
Moves P2P signaling instance initialization to the connect method where
the options are first accessible. This means that if the P2P signaling
instance is present then the P2P is enabled.

Also doesn't kill the P2P when Spot TV leaves the room. More logic is
required to handle P2P reconnects, but this will come as the next step.