ref(RCS+xmpp-connection): abort silent reconnect
If P2P goes down when the xmpp is in the silent reconnect process it
will not propagate the 'no-signaling' error and the user may think that
the connection is still alive, because the reconnecting label will not
be displayed. At least not until the silent reconnect either fails or
succeeds which may take a while in BOSH mode.