feat: detect Spot TV conflict
If Spot TV will fail to join the MUC with conflict error, for longer
than the max client idle timeout, it will mean that most likely another
Spot TV instance is already connected.

The conflict can not be detected instantly, because not gracefully
closed connections are cleaned up only after the idle timeout on
the server.