fix(share-mode.spec): adjust backend/no backend scenarios
When the backend is enabled temporary remote is disconnected after
a call and without the  backend it doesn't, so adjust the cases
accordingly.

Also fixes and re-enables the "Spot-Remote is disconnected on share end"
test. When remote is disconnect it lands on the join code page and
that's where the test should end.