ref(zoom): use a combined selector to calculate domain whitelist
This way multiple code paths can use the same logic to get
the full list of whitelisted domains,, which is now a generated
list due to zoom being whitelisted without being explicitly
specified as whitelisted in the config.

Reselect is brought in so that the generated array
of allowed domains can be memoized to avoid triggering
re-renders when the array is passed in as props.