fix(error-boundary): remove View component usage
The View component uses RCS to notify remotes
of the TV's current view. However, if an error
occurred in RCS, it is possible that trying to
use View will cause another error in RCS,
preventing the error boundary from displaying.
Instead, just show the status overlay and
let the TV reconnect.