fix(p2p): catch RTCDataChannel.send error
If send is called from the RTCDataChannel 'onmessage' event callback,
the channel's ready state or peerconnection's ICE state may not be up
to date and the app may crash while trying to send on closed connection.

Make sendDataChannelMessage return boolean and never throw. Handles
the result only where it matters.