feat: add support for the calendar push notifications
This will make the calendar service refresh the events when the push
notification is received. This is currently implemented only for
the backend calendar and the push mechanism is in the Prosody module.
Requires setting CALENDAR_PUSH_NOTIFICATIONS_ENABLED to true to make
the application aware.

With push enabled it is recommended to set the CALENDAR_POLLING_INTERVAL
to relatively high value as a backup. There's logic which emits
'calendar-push-missed' analytics event if new calendar events will be
found on the polling refresh. It means that the push notification has
been missed or not emitted properly by the backend.